In append_whiskers_to_hdf5, get follicle and tip coordinates from measurements data rather than whiskers data; also, convert MeasurementsTable to array using .as_array() instead of .get_shape_table() and use cxrodgers' interpretation of the column meanings
Get the whisker and follicle information from measurements information wheere possible; the reason for this is that the whiskers data itself doesn't encode which end of a whisker object is the follicle or the tip. cxrodgers made some assumtpions about which was which and these worked for him, but they don't generalize to other face orientations. whisk's measure function does the work of interpreting which end of the whisker is which based on the face orientation, so we might as well get this information from the measurements data where possible

also, convert MeasurementsTable to an array using .as_array() and use cxrodgers' interpretation of the columns, instead of converting using .get_shape_table(). For the latter, the interpreation of the columns was different, and I don't remember where they came from, so I'd rather convert the MeasurementsTable into a format where there's some documentation of how to interpret the columns (although evidently this inerpretation comes from cxrodgers' own experimentation with the output, rather than Nathan Clack's own documentation, which nobody can find). The new interpreation of the columns can be found in the comments for base.py in https://github.com/cxrodgers/whiskvid.git/